requirements.txt
django==4.0.0
djangorestframework==3.12.4

django-filter==2.4.0
django-cors-headers==3.7.0
django-extensions==3.1.3

redgreenunittest==0.1.1
drf-spectacular==0.25.0

python-dotenv~=0.21.0
pytz==2022.7.1

psycopg2-binary==2.9.5
urllib3~=1.26.13
mysqlclient==2.1.1
django-mysql==3.8.1

dj-rest-auth==4.0.1
django-allauth==0.50.0
djangorestframework-simplejwt==5.2.2

PyJWT==2.7.0
